+++ src/routes/(v2)/v2/Navigation.svelte
@@ -72,7 +72,8 @@
 		'arktype',
 		'effect',
 		'issue-500',
-		'typebox'
+		'typebox',
+		'transport'
 	].sort();
 </script>